requirements.txt
# Core dependencies
python-dotenv==1.0.0
toml==0.10.2
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==1.0.1

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Web scraping and parsing
lxml==4.9.3
html5lib==1.1
chardet==5.2.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1 